Added parameters to scripts
